docs: Describe project goals
Following a conversation [here], this commit tries to document
project goals.  We can use these goals as a compass to decide which
changes are in scope and which are out of scope.

Currently, this project is meant to be a lightweight layer for
interacting with Bybit's WebSocket API.  Performance is king, and this
means many conveniences may have to be built into separate packages
that build upon this package.

This should make it easier to bring this project into a "complete"
state.

[here]: https://github.com/yufuquant/rust-bybit/pull/8#issuecomment-1182013219